requirements.txt
requests
pydub
opencv-python
numpy
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
gspread
torch
TTS
cairosvg
Pillow
maturin
